Move validation to different module to prevent confusion
